Verify the status code of the returned data before passing it into a function for execution.
Many sites that are misconfigured or unexpectedly down will return an HTML page along with a status code of 401, 404, 405, or in YQL's case, 999. Rather than assume the Content-Type is correctly set on the return, check the status code before passing the returned body into loaderFn to become a new function that's immediately executed. Doing so with passed-in HTML formatted text will result in syntax errors that cannot be caught and will halt execution.